refactor(file_ops): restructure parser system following hexagonal architecture
