+++ fileorg/file_ops/adapters/file_parser/csv_parser.py
@@ -9,7 +9,7 @@ from pathlib import Path
 
 import chardet
 
-from fileorg.file_ops.ports import IParser, ParserOutput
+from fileorg.file_ops.ports.parser_ports import IParser, ParserOutput
 
 
 class CsvParser(IParser):
