+++ fileorg/file_ops/adapters/file_parser/html_parser.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import chardet
 from bs4 import BeautifulSoup
 
-from fileorg.file_ops.ports import IParser, ParserOutput
+from fileorg.file_ops.ports.parser_ports import IParser, ParserOutput
 
 
 class HtmlParser(IParser):
