+++ fileorg/file_ops/adapters/file_parser/ppt_parser.py
@@ -1,7 +1,7 @@
 # ppt_parser.py
 from pathlib import Path
 
-from fileorg.file_ops.ports import IParser, ParserOutput
+from fileorg.file_ops.ports.parser_ports import IParser, ParserOutput
 
 
 class PptParser(IParser):
