+++ fileorg/file_ops/adapters/file_parser/txt_parser.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import chardet
 
-from fileorg.file_ops.ports import IParser, ParserOutput
+from fileorg.file_ops.ports.parser_ports import IParser, ParserOutput
 
 
 class TxtParser(IParser):
